requirements.txt
python-dotenv>=1.0.1
pydantic>=2.6.0
langchain>=0.2.12
langchain-openai>=0.1.22
langgraph>=0.2.0
requests>=2.31.0
